implement filtering on pool-overview for bureau users, refactor some code into its own file for better maintenance, refactor some daos into the new format
