+++ client/templates/pool-management/pool-overview/_partials/filters.njk
@@ -5,156 +5,166 @@
 
 {% set filterOptionsHtml %}
 
-{{ govukInput({
-  label: {
-    text: "Juror number"
-  },
-  id: "jurorNumber",
-  name: "jurorNumber",
-  value: appliedFilters.jurorNumber
-}) }}
+  {{ govukInput({
+    label: {
+      text: "Juror number"
+    },
+    id: "jurorNumber",
+    name: "jurorNumber",
+    value: appliedFilters.jurorNumber
+  }) }}
 
-{{ govukInput({
-  label: {
-    text: "First name"
-  },
-  id: "firstName",
-  name: "firstName",
-  value: appliedFilters.firstName
-}) }}
+  {{ govukInput({
+    label: {
+      text: "First name"
+    },
+    id: "firstName",
+    name: "firstName",
+    value: appliedFilters.firstName
+  }) }}
 
-{{ govukInput({
-  label: {
-    text: "Last name"
-  },
-  id: "lastName",
-  name: "lastName",
-  value: appliedFilters.lastName
-}) }}
+  {{ govukInput({
+    label: {
+      text: "Last name"
+    },
+    id: "lastName",
+    name: "lastName",
+    value: appliedFilters.lastName
+  }) }}
 
-{{ govukCheckboxes({
-  name: "attendance",
-  classes: "govuk-checkboxes--small",
-  fieldset: {
-    legend: {
-      text: "Attendance",
-      classes: "govuk-fieldset__legend--s"
-    }
-  },
-  values: appliedFilters.attendance,
-  items: [
-    {
-      value: "in attendance",
-      text: "In attendance"
-    },
-    {
-      value: "on call",
-      text: "On call"
-    },
-    {
-      value: "other",
-      text: "Other"
-    },
-    {
-      value: "on a trial",
-      text: "On a trial"
-    }
-  ]
-}) }}
+  {% if isCourtUser %}
+    {{ govukCheckboxes({
+      name: "attendance",
+      classes: "govuk-checkboxes--small",
+      fieldset: {
+        legend: {
+          text: "Attendance",
+          classes: "govuk-fieldset__legend--s"
+        }
+      },
+      values: appliedFilters.attendance,
+      items: [
+        {
+          value: "in attendance",
+          text: "In attendance"
+        },
+        {
+          value: "on call",
+          text: "On call"
+        },
+        {
+          value: "other",
+          text: "Other"
+        },
+        {
+          value: "on a trial",
+          text: "On a trial"
+        }
+      ]
+    }) }}
 
-{{ govukCheckboxes({
-  name: "checkedIn",
-  classes: "govuk-checkboxes--small",
-  fieldset: {
-    legend: {
-      text: "Checked in",
-      classes: "govuk-fieldset__legend--s"
-    }
-  },
-  values: appliedFilters.checkedIn,
-  items: [
-    {
-      value: "today",
-      text: "Checked in today"
-    }
-  ]
-}) }}
+    {{ govukCheckboxes({
+      name: "checkedIn",
+      classes: "govuk-checkboxes--small",
+      fieldset: {
+        legend: {
+          text: "Checked in",
+          classes: "govuk-fieldset__legend--s"
+        }
+      },
+      values: appliedFilters.checkedIn,
+      items: [
+        {
+          value: "today",
+          text: "Checked in today"
+        }
+      ]
+    }) }}
 
-{{ govukCheckboxes({
-  name: "nextDueAtCourt",
-  classes: "govuk-checkboxes--small",
-  fieldset: {
-    legend: {
-      text: "Next due at court",
-      classes: "govuk-fieldset__legend--s"
-    }
-  },
-  values: appliedFilters.nextDueAtCourt,
-  items: [
-    {
-      value: "set",
-      text: "Date set"
-    },
-    {
-      value: "notSet",
-      text: "No date set"
-    }
-  ]
-}) }}
+    {{ govukCheckboxes({
+      name: "nextDueAtCourt",
+      classes: "govuk-checkboxes--small",
+      fieldset: {
+        legend: {
+          text: "Next due at court",
+          classes: "govuk-fieldset__legend--s"
+        }
+      },
+      values: appliedFilters.nextDueAtCourt,
+      items: [
+        {
+          value: "set",
+          text: "Date set"
+        },
+        {
+          value: "notSet",
+          text: "No date set"
+        }
+      ]
+    }) }}
+  {% endif %}
 
-{{ govukCheckboxes({
-  name: "status",
-  classes: "govuk-checkboxes--small",
-  fieldset: {
-    legend: {
-      text: "Status",
-      classes: "govuk-fieldset__legend--s"
-    }
-  },
-  values: appliedFilters.status,
-  items: [
-    {
-      value: "responded",
-      text: "Responded"
+  {{ govukCheckboxes({
+    name: "status",
+    classes: "govuk-checkboxes--small",
+    fieldset: {
+      legend: {
+        text: "Status",
+        classes: "govuk-fieldset__legend--s"
+      }
     },
-    {
-      value: "panel",
-      text: "Panelled"
-    },
-    {
-      value: "juror",
-      text: "Juror"
-    },
-    {
-      value: "completed",
-      text: "Completed"
-    },
-    {
-      value: "FailedToAttend",
-      text: "Failed to attend"
-    },
-    {
-      value: "deferred",
-      text: "Deferred"
-    },
-    {
-      value: "excused",
-      text: "Excused"
-    },
-    {
-      value: "postponed",
-      text: "Postponed"
-    },
-    {
-      value: "disqualified",
-      text: "Disqualified"
-},
-    {
-      value: "deceased",
-      text: "Deceased"
-    }
-  ]
-}) }}
+    values: appliedFilters.status,
+    items: [
+      {
+        value: "summoned",
+        text: "Summoned"
+      } if isBureauUser,
+      {
+        value: "responded",
+        text: "Responded"
+      },
+      {
+        value: "panel",
+        text: "Panelled"
+      } if isCourtUser,
+      {
+        value: "juror",
+        text: "Juror"
+      } if isCourtUser,
+      {
+        value: "completed",
+        text: "Completed"
+      } if isCourtUser,
+      {
+        value: "FailedToAttend",
+        text: "Failed to attend"
+      } if isCourtUser,
+      {
+        value: "deferred",
+        text: "Deferred"
+      },
+      {
+        value: "excused",
+        text: "Excused"
+      },
+      {
+        value: "postponed",
+        text: "Postponed"
+      },
+      {
+        value: "disqualified",
+        text: "Disqualified"
+      },
+      {
+        value: "deceased",
+        text: "Deceased"
+      },
+      {
+        value: "undeliverable",
+        text: "Undeliverable"
+      } if isBureauUser
+    ]
+  }) }}
 
 {% endset %}
 
@@ -163,7 +173,7 @@
     heading: {
       text: "Filter"
     },
-classes: "js-hidden",
+    classes: "js-hidden",
     attributes: {
       id: "pool-overview-filter",
       "aria-hidden": true
