+++ client/templates/pool-management/pool-overview/bureau-pool-jurors.njk
@@ -47,9 +47,10 @@
         text: "Show filter",
         classes: "govuk-button--secondary govuk-!-margin-bottom-0",
         attributes: {
-          id: "editPoolButton"
+          id: "pool-overview-show-filter"
         }
       }) }}
+
       {{ govukButton({
         text: "Reassign",
         classes: "govuk-button--secondary govuk-!-margin-bottom-0",
@@ -59,6 +60,7 @@
           formAction: postUrls.assignUrl
         }
       }) }}
+
       {{ govukButton({
         text: "Postpone",
         classes: "govuk-button--secondary govuk-!-margin-bottom-0",
@@ -68,7 +70,9 @@
           formAction: postUrls.postponeUrl
         }
       }) }}
+
       {% if isCourtUser %}
+
         {{ govukButton({
           text: "Transfer",
           classes: "govuk-button--secondary govuk-!-margin-bottom-0",
@@ -78,6 +82,7 @@
             formAction: postUrls.transferUrl
           }
         }) }}
+
         {{ govukButton({
           text: "Complete service",
           classes: "govuk-button--secondary govuk-!-margin-bottom-0",
@@ -87,50 +92,70 @@
             formAction: postUrls.completeServiceUrl
           }
         }) }}
+
       {% endif %}
     </div>
-{% if totalJurors === 0 and poolDetails.isActive === true %}
 
-  <p class="govuk-body">
-    This pool has no members.
-  </p>
+    <div class="govuk-grid-row">
 
-  {% else %}
-     {% if totalJurors > 500 %}
-        <span class="govuk-body">There are more than 500 results. Refine your search.</span>
-    {% else %} 
-      {% if errors.items["selectedJurors"] %}<div class="error-allocation">
-        <p class="govuk-error-message">{{ errors.items['selectedJurors'][0].details }}</p>
+      <div class="govuk-grid-column-one-third">
+        {% include "./_partials/filters.njk" %}
+      </div>
+
+      {% if totalJurors > 500 %}
+
+        <div id="pool-overview-table-wrapper" class="govuk-grid-column-full">
+          <span class="govuk-body">There are more than 500 results. Refine your search.</span>
+        </div>
+
+      {% elif totalJurors === 0 %}
+
+        <div id="pool-overview-table-wrapper" class="govuk-grid-column-full">
+          <span class="govuk-body">There are no results to display.</span>
+        </div>
+
+      {% else %}
+
+        <div id="pool-overview-table-wrapper" class="govuk-grid-column-full">
+
+          {% if errors.items["selectedJurors"] %}<div class="error-allocation">
+            <p class="govuk-error-message">{{ errors.items['selectedJurors'][0].details }}</p>
+          {% endif %}
+        
+          {% if errors.items["selectedJurors"] %}<div class="govuk-form-group govuk-form-group--error">{% endif %}
+
+          <div class="govuk-caption-m">
+            <span id="total-checked-jurors">{{ totalCheckedJurors }}</span> of <span id="total-jurors-count">{{ totalJurors }}</span> selected
+          </div>
+
+          {{ modSortableTable({
+            id: "jurorOverview",
+            caption: "Jurors in pool",
+            captionClass: "govuk-visually-hidden",
+            head: membersHeaders,
+            rows: poolMembers,
+            url: urlPrefix
+          }) }}
+
+          {% if errors.items["selectedJurors"] %}</div>{% endif %}
+
+          {{ govukPagination({
+            previous: {
+              href: pageItems.prev
+            },
+            next: {
+              href: pageItems.next
+            },
+            items: pageItems.items
+          }) }}
+
+          {% if errors.items["selectedJurors"] %}</div>{% endif %}
+
+        </div>
+
       {% endif %}
-      
-      {% if errors.items["selectedJurors"] %}<div class="govuk-form-group govuk-form-group--error">{% endif %}
 
-        <div class="govuk-body"> <span id="total-checked-jurors">{{ totalCheckedJurors }}</span> of <span id="total-jurors-count">{{ totalJurors }}</span> selected</div>
-
-        {{ modSortableTable({
-          id: "jurorOverview",
-          caption: "Jurors in pool",
-          captionClass: "govuk-visually-hidden",
-          head: membersHeaders,
-          rows: poolMembers,
-          url: urlPrefix
-        }) }}
-
-        {% if errors.items["selectedJurors"] %}</div>{% endif %}
-
-        {{ govukPagination({
-          previous: {
-            href: pageItems.prev
-          },
-          next: {
-            href: pageItems.next
-          },
-          items: pageItems.items
-        }) }}
-
- {% endif %}
-      {% if errors.items["selectedJurors"] %}</div>{% endif %}
-    {% endif %}
+    </div>
 
     <input type="hidden" name="navigate" id="navigate" value="">
     <input type="hidden" name="poolNumber" id="poolNumber" value="{{ poolDetails.poolNumber }}">
